не получается сделать диалог задержки через liveData

+++ tempo_leader_javanew/src/main/java/ru/barcats/tempo_leader_javanew/ui/sekundomer/grafic/GraficStorageImpl.java
@@ -85,7 +85,7 @@ public class GraficStorageImpl implements GraficStorage {
     }
 
     @Override
-    public long getColumnIndex(Cursor cursor) {
+    public int getColumnIndex(Cursor cursor) {
         return cursor.getColumnIndex(TabSet.COLUMN_SET_TIME);
     }
